Configure integrated env using docker-compose
The lack of centralized integration of services for development has been a
difficulty since the start of this project. We would need to configure services
between different projects, resulting in a scattered configuration all across
the file system for things that should interact.

By introducing a structured set of docker-compose configurations, based on the
repositories' names, along with scripts to gather all of them and integrate on a
shared network, we're able to create a local system with all the services
communicating through the overlay that docker provides.