Remove docker in favor of kind cluster
As there'll be no shared network over docker and kind creates a cluster
with the required connectivity between services, we can ignore that on
the docker layer and leave it to the Kubernetes cluster to handle.

Also, the centralized configuration is being discouraged in favor of per
repository configs, thus, removing the need for targets approach and
anything that's project-specific inside this repository. From now on,
this project will be only responsible for setting up a local environment
where developers submit deployments to validate their code before
submitting to the respective upstream.