Configure k8s components to enable HTTPS traffic
For our project, an environment without HTTPS would give us a
handicapped experience once the browser blocks us from accessing some
APIs.

Therefore, we need to be able to route external traffic to our services
through a secure channel, which is handled by [Contour][contour], as our
Ingress Controller that designates external traffic, and
[Cert-Manager][cert-manager] to generate certificates based on a root
certificate.

Contour was the chosen one for the ingress controller because of its
`IngressRoute` CRD that enables a decentralized ingress configuration
complying with the project direction; also, we choose it because of
existing guides that have shown the ease of configuring Contour inside
Kind. For the TLS side, Cert-Manager was a clear choice given its spread
on the community despite the API instability.

To glue everything together, we created resources needed to have a
`ClusterIssuer` pointing to the secret that holds a root certificate
used for issuing a new certificate. I've chosen to use `ClusterIssuer`
due to reach over all the namespaces on the cluster. Regarding the
certificate, I decided to use the ca issuer, instead of just a
self-signed issuer, because we need to trust a root certificate to
prevent the browser from distrusting our local environment.

[contour]: https://projectcontour.io/
[cert-manager]: https://cert-manager.io/