Embed the docker registry inside the cluster
Having a completely separate container just for the docker registry for
local development is not a pleasant experience. Because of this
arrangement, the cluster can't be stopped as a whole if we consider the
registry a part of it. Besides that, having to push images to
`localhost:5000` instead of the target we'd normally be pushing
to (`registry.oddin.org`) is a little troublesome.

Therefore, we configured the registry with a Pod within the cluster
itself. To access the registry, we used a Service with type `NodePort`
on `30008`, enabling the use of `localhost:30008` as a mirror to
`registry.oddin.org`.