Add Ingress TLS placeholder for HTTPs with Contour
As HTTPProxy from Contour doesn't integrate directly with Cert-Manager
ways of working, we need to create a valid TLS secret to be used by our
proxies. The responsible for the creation is a dummy ingress that acts
only as an integration point to Cert-Manager.

It seems strange to have an Ingress along with HTTPProxy with both
mapping the same domain, and you might think that they will clash.
However, since the Ingress doesn't have the `ingress-class: contour`
annotation, Contour doesn't route through the Ingress and only uses the
HTTPProxy.